Refactoring controller functions Creating authToken file for generating token for authentications Creating service for database queries
