+++ controllers/userController.js
@@ -1,33 +1,11 @@
 const { promisify } = require('util');
 const bcrypt = require('bcryptjs');
 const db = require('../knex/knex');
-const jwt = require('jsonwebtoken');
-const dotenv = require('dotenv');
+const { createSendToken } = require('../utils/authToken');
 const catchAsync = require('../utils/catchAsync');
 const AppError = require('../utils/appError');
 const hashedPass = require('../utils/hashedPassword');
-
-dotenv.config({ path: './config.env' });
-
-const signToken = (id) => {
-  return jwt.sign({ id }, process.env.JWT_SECRET, {
-    expiresIn: process.env.JWT_EXPIRES_IN,
-  });
-};
-
-const createSendToken = (user, statusCode, res) => {
-  const token = signToken(user);
-  // user.password = undefined;
-  // user.password_confirm = undefined;
-
-  res.status(statusCode).json({
-    status: 'success',
-    token,
-    data: {
-      user,
-    },
-  });
-};
+const { saveUser, logUser } = require('../service/userService');
 
 exports.createUser = catchAsync(async (req, res, next) => {
   let { email, first_name, last_name, password, password_confirm } = req.body;
@@ -37,34 +15,20 @@ exports.createUser = catchAsync(async (req, res, next) => {
     );
   }
   const hashedPassword = await hashedPass(password);
-  // password_confirm = undefined;
-  const newUser = await db('users').insert({
-    email,
-    first_name,
-    last_name,
-    password: hashedPassword,
-  });
+  const newUser = await saveUser(email, first_name, last_name, hashedPassword);
   createSendToken(newUser[0], 201, res);
-  // } catch (err) {
-  //   console.log(err);
 });
 
 exports.login = catchAsync(async (req, res, next) => {
   const { email, password } = req.body;
-  //Check if email and password exist
   if (!email || !password) {
-    // throw new Error('Please provide email and password');
     return next(new AppError('Please provide email or password', 404));
   }
-  //Check if user exist and password is correct
-  const user = (await db('users').where('email', email))[0];
-  // const hashedPassword = await hashedPass(password)
+  const user = await logUser(email);
 
   if (!user || !(await bcrypt.compare(password, user.password))) {
-    // throw new Error('Wrong user name or password')
     return next(new AppError('Wrong user name or password', 404));
   }
-  //if everything is ok, send token to client
   createSendToken(user.id, 200, res);
 });
 
@@ -79,19 +43,18 @@ exports.deposit = catchAsync(async (req, res, next) => {
     return next(new AppError('invalid amount', 400));
   }
   const newWallet = (user[0].wallet += amount);
-  // await db('users').insert()
   await db('users')
     .where({ email: req.body.email })
     .update('wallet', newWallet);
-  // res.status(200).json({
-  //   status: 'success',
-  //   user,
-  // });
-  createSendToken(user[0].id, 200, res);
+
+  // createSendToken(user[0].id, 200, res);
+  res.status(200).json({
+    status: 'success',
+    message:`${amount} successfully deposited}`
+  })
 });
 
 exports.protect = catchAsync(async (req, res, next) => {
-  //Get the token and check if it is there
   let token;
   if (
     req.headers.authorization &&
@@ -105,26 +68,16 @@ exports.protect = catchAsync(async (req, res, next) => {
       new AppError('You are not logged in! Please log in to get access', 401)
     );
   }
-  //Token verification
+
   const decoded = await promisify(jwt.verify)(token, process.env.JWT_SECRET);
-  //Check if user still exists
-  // console.log(decoded);
   const currentUser = await db('users').where('id', decoded.id);
+
   if (!currentUser) {
     return next(
       new AppError('The user belonging to this  token no longer exist', 401)
     );
   }
-  //Check if user changed password after the token was issued
-  // if (await currentUser.changedPasswordAfter(decoded.iat)) {
-  //   return next(
-  //     new AppError('User recently changed password! Please log in again', 401)
-  //   );
-  // }
-  // console.log(await currentUser.changedPasswordAfter(decoded.iat));
 
-  //Grant access to protected route
-  // console.log(currentUser);
   req.user = currentUser;
 
   next();
@@ -150,7 +103,6 @@ exports.withdraw = catchAsync(async (req, res, next) => {
     message: `${amount} sucessfully witdrawn`,
     walletBalance: `${wallet}`,
   });
-  // createSendToken(user.id, 200, res);
 });
 
 exports.transfer = catchAsync(async (req, res, next) => {
